GarminHUD: Move to sky4s.garminhud.hud namespace
* Since this is a HUD class, move it to the HUD package.
